+++ finality-provider/cmd/fpd/daemon/commit_pr.go
@@ -105,8 +105,13 @@ func runCommandCommitPubRand(ctx client.Context, cmd *cobra.Command, args []stri
 	fpMetrics := metrics.NewFpMetrics()
 	poller := service.NewChainPoller(logger, cfg.PollerConfig, consumerCon, fpMetrics)
 
+	rndCommitter := service.NewDefaultRandomnessCommitter(
+		service.NewRandomnessCommitterConfig(cfg.NumPubRand, int64(cfg.TimestampingDelayBlocks), cfg.ContextSigningHeight),
+		service.NewPubRandState(pubRandStore), consumerCon, em, logger, fpMetrics)
+	heightDeterminer := service.NewStartHeightDeterminer(consumerCon, cfg.PollerConfig, logger)
+
 	fp, err := service.NewFinalityProviderInstance(
-		fpPk, cfg, fpStore, pubRandStore, cc, consumerCon, em, poller, fpMetrics,
+		fpPk, cfg, fpStore, pubRandStore, cc, consumerCon, em, poller, rndCommitter, heightDeterminer, fpMetrics,
 		make(chan<- *service.CriticalError), logger)
 	if err != nil {
 		return fmt.Errorf("failed to create finality-provider %s instance: %w", fpPk.MarshalHex(), err)