+++ finality-provider/service/app.go
@@ -34,14 +34,16 @@ type FinalityProviderApp struct {
 	wg        sync.WaitGroup
 	quit      chan struct{}
 
-	cc           ccapi.ClientController
-	consumerCon  ccapi.ConsumerController
-	kr           keyring.Keyring
-	fps          *store.FinalityProviderStore
-	pubRandStore *store.PubRandProofStore
-	config       *fpcfg.Config
-	logger       *zap.Logger
-	poller       types.BlockPoller[types.BlockDescription]
+	cc               ccapi.ClientController
+	consumerCon      ccapi.ConsumerController
+	kr               keyring.Keyring
+	fps              *store.FinalityProviderStore
+	pubRandStore     *store.PubRandProofStore
+	config           *fpcfg.Config
+	logger           *zap.Logger
+	poller           types.BlockPoller[types.BlockDescription]
+	rndCommitter     types.RandomnessCommitter
+	heightDeterminer types.HeightDeterminer
 
 	fpInsMu     sync.RWMutex // Protects fpIns
 	fpIns       *FinalityProviderInstance
@@ -85,7 +87,23 @@ func NewFinalityProviderAppFromConfig(
 	fpMetrics := metrics.NewFpMetrics()
 	poller := NewChainPoller(logger, cfg.PollerConfig, consumerCon, fpMetrics)
 
-	return NewFinalityProviderApp(cfg, cc, consumerCon, em, poller, fpMetrics, db, logger)
+	pubRandStore, err := store.NewPubRandProofStore(db)
+	if err != nil {
+		return nil, fmt.Errorf("failed to initiate public randomness store: %w", err)
+	}
+
+	rndCommiter := NewDefaultRandomnessCommitter(
+		NewRandomnessCommitterConfig(cfg.NumPubRand, int64(cfg.TimestampingDelayBlocks), cfg.ContextSigningHeight),
+		NewPubRandState(pubRandStore),
+		consumerCon,
+		em,
+		logger,
+		fpMetrics,
+	)
+
+	heightDeterminer := NewStartHeightDeterminer(consumerCon, cfg.PollerConfig, logger)
+
+	return NewFinalityProviderApp(cfg, cc, consumerCon, em, poller, rndCommiter, heightDeterminer, fpMetrics, db, logger)
 }
 
 func NewFinalityProviderApp(
@@ -94,6 +112,8 @@ func NewFinalityProviderApp(
 	consumerCon ccapi.ConsumerController,
 	em eotsmanager.EOTSManager,
 	poller types.BlockPoller[types.BlockDescription],
+	rndCommitter types.RandomnessCommitter,
+	heightDeterminer types.HeightDeterminer,
 	metrics *metrics.FpMetrics,
 	db kvdb.Backend,
 	logger *zap.Logger,
@@ -126,6 +146,8 @@ func NewFinalityProviderApp(
 		logger:                            logger,
 		eotsManager:                       em,
 		poller:                            poller,
+		rndCommitter:                      rndCommitter,
+		heightDeterminer:                  heightDeterminer,
 		metrics:                           metrics,
 		quit:                              make(chan struct{}),
 		unjailFinalityProviderRequestChan: make(chan *UnjailFinalityProviderRequest),
@@ -538,7 +560,8 @@ func (app *FinalityProviderApp) startFinalityProviderInstance(
 
 	if app.fpIns == nil {
 		fpIns, err := NewFinalityProviderInstance(
-			pk, app.config, app.fps, app.pubRandStore, app.cc, app.consumerCon, app.eotsManager, app.poller,
+			pk, app.config, app.fps, app.pubRandStore, app.cc, app.consumerCon,
+			app.eotsManager, app.poller, app.rndCommitter, app.heightDeterminer,
 			app.metrics, app.criticalErrChan, app.logger,
 		)
 		if err != nil {