+++ finality-provider/service/pub_rand_store_adapter.go
@@ -5,29 +5,29 @@ import (
 	"github.com/cometbft/cometbft/crypto/merkle"
 )
 
-type pubRandState struct {
+type PubRandState struct {
 	s *store.PubRandProofStore
 }
 
-func newPubRandState(s *store.PubRandProofStore) *pubRandState {
-	return &pubRandState{s: s}
+func NewPubRandState(s *store.PubRandProofStore) *PubRandState {
+	return &PubRandState{s: s}
 }
 
-func (st *pubRandState) addPubRandProofList(
+func (st *PubRandState) addPubRandProofList(
 	pk, chainID []byte, height uint64, numPubRand uint64,
 	proofList []*merkle.Proof,
 ) error {
 	return st.s.AddPubRandProofList(chainID, pk, height, numPubRand, proofList)
 }
 
-func (st *pubRandState) getPubRandProof(pk, chainID []byte, height uint64) ([]byte, error) {
+func (st *PubRandState) getPubRandProof(pk, chainID []byte, height uint64) ([]byte, error) {
 	return st.s.GetPubRandProof(chainID, pk, height)
 }
 
-func (st *pubRandState) getPubRandProofList(pk, chainID []byte, height uint64, numPubRand uint64) ([][]byte, error) {
+func (st *PubRandState) getPubRandProofList(pk, chainID []byte, height uint64, numPubRand uint64) ([][]byte, error) {
 	return st.s.GetPubRandProofList(chainID, pk, height, numPubRand)
 }
 
-func (st *pubRandState) close() error {
+func (st *PubRandState) close() error {
 	return st.s.Close()
 }